creation des differents fichiers de screens, initialisations des tables de la DB
