+++ app/src/main/java/com/example/lacasa/data/Building.kt
@@ -1,0 +1,3 @@
+package com.example.lacasa.data
+
+data class Building()
