+++ app/src/main/java/com/example/lacasa/data/Tenant.kt
@@ -1,0 +1,3 @@
+package com.example.lacasa.data
+
+data class Tenant()
