+++ app/src/main/java/com/example/lacasa/model/BuildingDao.kt
@@ -1,0 +1,4 @@
+package com.example.lacasa.model
+
+interface BuildingDao {
+}
