+++ app/src/main/java/com/example/lacasa/model/LocalDateConverter.kt
@@ -1,0 +1,24 @@
+package com.example.todolist.model
+
+import android.os.Build
+import androidx.annotation.RequiresApi
+import androidx.room.TypeConverter
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+object LocalDateConverter {
+    @RequiresApi(Build.VERSION_CODES.O)
+    private val formatter = DateTimeFormatter.ofPattern(/* pattern = */ "yyyy-MM-dd")
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    @TypeConverter
+    fun fromLocalDate(date: LocalDate): String {
+        return date.format(formatter)
+    }
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    @TypeConverter
+    fun toLocalDate(dateString: String): LocalDate {
+        return LocalDate.parse(dateString, formatter)
+    }
+}
