+++ app/src/main/java/com/example/lacasa/model/PaymentDao.kt
@@ -1,0 +1,4 @@
+package com.example.lacasa.model
+
+interface PaymentDao {
+}
