+++ app/src/main/java/com/example/lacasa/model/TaskDao.kt
@@ -1,0 +1,26 @@
+package com.example.todolist.model
+
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.Query
+import androidx.room.Update
+
+import androidx.room.OnConflictStrategy
+import com.example.todolist.data.Task
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface TaskDao {
+    @Query("SELECT * FROM task ORDER BY dateFin ASC")
+    fun getAllTasks():Flow<List<Task>>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertTask(task:Task)
+
+    @Update
+    suspend fun updateTask(task:Task)
+
+    @Delete
+    suspend fun deleteTask(task:Task)
+}
