+++ app/src/main/java/com/example/lacasa/model/TenantDao.kt
@@ -1,0 +1,4 @@
+package com.example.lacasa.model
+
+interface TenantDao {
+}
