+++ app/src/main/java/com/example/lacasa/model/UserDao.kt
@@ -1,0 +1,4 @@
+package com.example.lacasa.model
+
+class UserDao {
+}
