+++ app/src/main/java/com/example/lacasa/ui/TenantScreen.kt
@@ -1,0 +1,4 @@
+package com.example.lacasa.ui
+
+class TenantScreen {
+}
