+++ gradle/libs.versions.toml
@@ -9,6 +9,14 @@ lifecycleRuntimeKtx = "2.8.7"
 activityCompose = "1.10.0"
 composeBom = "2024.04.01"
 
+#Navigation
+lifecycle-viewmodel-compose = "2.8.7"
+navigation-compose="2.8.6"
+
+#BDD | Room
+android-room = "2.6.1"
+ksp = "2.0.21-1.0.27"
+
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
@@ -25,8 +33,21 @@ androidx-ui-test-manifest = { group = "androidx.compose.ui", name = "ui-test-man
 androidx-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit4" }
 androidx-material3 = { group = "androidx.compose.material3", name = "material3" }
 
+#Navigation
+androidx-lifecycle-viewmodel= { group = "androidx.lifecycle", name = "lifecycle-viewmodel-compose", version.ref = "lifecycle-viewmodel-compose" }
+androidx-navigation={ group = "androidx.navigation", name = "navigation-compose", version.ref = "navigation-compose"}
+
+#BDD | Room
+androidx-room={ group = "androidx.room", name = "room-ktx", version.ref = "android-room" }
+androidx-room-runtime={ group = "androidx.room", name = "room-runtime", version.ref = "android-room" }
+androidx-room-compiler={ group = "androidx.room", name = "room-compiler", version.ref = "android-room" }
+
+
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
+
+#ksp
+google-devtools = { id = "com.google.devtools.ksp", version.ref = "ksp" }
